Conflict resolutions favor main's production code and keep this
branch's test repairs adapted to it:
- state/world.rs: main's entity_creation_order_pairs adapter supersedes
  the any-key map adapter (same bug, fixed independently on both sides).
- sim_rng.rs: both sides converged on ScriptRand=2 and the
  script_rand_with_context entry point; main's more precise
  macro-token site filter replaces the string-replace variant.
- engine/mod.rs: keep both fixture semantics — main's always-provision
  profiles plus this branch's campaign-description identity attachment
  (pcs_mut retained for the identity write-back).
- tests/{commands,macros,movement}.rs: main's test bodies (paired with
  its movement parity changes) with this branch's comment where it adds
  context.

+++ crates/robin_engine/src/natives/dispatch/ai.rs
@@ -588,8 +588,11 @@ impl NativeContext<'_, '_> {
             ForbidNPCRemark => {
                 // ForbidNPCRemark(Actor, int remark_id, bool forbid)
                 // Adds or removes a remark ID from this NPC's forbidden list.
-                let forbid = stack.pop_i32();
-                let remark_id = stack.pop_i32();
+                // Both trailing arguments are narrowed to a signed byte before
+                // they reach the implementation, the same way
+                // `SetPersistentProperty` narrows its own.
+                let forbid = i32::from(stack.pop_i32() as i8);
+                let remark_id = i32::from(stack.pop_i32() as i8);
                 let actor = stack.pop_i32();
                 if let Some(entity) = self.get_entity_mut(actor)
                     && let Some(ai) = entity.ai_controller_mut()
